Adding a 2nd property on the NanoXML benchmark
