+++ java/java-ranger-regression/nanoxml_eqchk/impl/ContentReader.java
